did some renaming addapted fragment_info.xml for code consistency
