+++ app/src/main/res/layout/frag_sustainability_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.info.SustainabilityListFrag">
+    tools:context=".ui.Sustainability.SustainabilityListFrag">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_main"
