+++ app/src/main/res/layout/fragment_info.xml
@@ -1,84 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.events.EventsFragment">
-
-    <Spinner
-        android:id="@+id/spinner"
-        android:layout_width="409dp"
-        android:layout_height="53dp"
-        android:layout_marginTop="56dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:ignore="SpeakableTextPresentCheck" />
+    android:layout_height="wrap_content"
+    tools:context=".ui.clubs.ClubListFragment" >
 
     <com.google.android.material.card.MaterialCardView
-        android:id="@+id/card"
-        android:clickable="true"
-        android:focusable="true"
+        android:id="@+id/cardview_home_sustainability"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_margin="20dp"
+        android:clickable="false"
         app:cardCornerRadius="15dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/spinner">
-
-        <LinearLayout
-
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-
-            <!-- Media -->
-
-            <ImageView
-                android:layout_width="match_parent"
-                android:layout_height="194dp"
-                android:scaleType="centerCrop"
-                android:src="@drawable/wetlands" />
-
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-
-                android:orientation="horizontal"
-                android:padding="12dp">
-
-                <!-- Title, secondary and supporting text -->
-
-                <!--android:textAppearance="?attr/textAppearanceTitleMedium" -->
-
-
-                <TextView
-                    android:id="@+id/textViewCampusInfo1"
-                    android:layout_width="240dp"
-                    android:layout_height="20dp"
-                    android:text="WETLANDS"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-
-            </androidx.constraintlayout.widget.ConstraintLayout>
-        </LinearLayout>
-
-    </com.google.android.material.card.MaterialCardView>
-
-    <com.google.android.material.card.MaterialCardView
-        android:id="@+id/card2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp"
-        app:cardCornerRadius="15dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/card">
-
+        app:layout_constraintTop_toTopOf="parent">
 
         <LinearLayout
 
@@ -89,10 +26,12 @@
             <!-- Media -->
 
             <ImageView
+                android:id="@+id/imageview_sustainability_banner"
                 android:layout_width="match_parent"
                 android:layout_height="194dp"
-                android:scaleType="centerCrop"
-                android:src="@drawable/green_buildings" />
+                android:contentDescription="@string/image_sustainability_banner"
+                android:scaleType="fitCenter"
+                android:src="@drawable/default_image" />
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
@@ -101,22 +40,20 @@
                 android:orientation="horizontal"
                 android:padding="12dp">
 
-                <!-- Title, secondary and supporting text -->
-
-                <!--android:textAppearance="?attr/textAppearanceTitleMedium" -->
-
-
                 <TextView
-                    android:id="@+id/textViewCampusInfo2"
-                    android:layout_width="240dp"
-                    android:layout_height="20dp"
-                    android:text="GREEN BUILDINGS"
-                    app:layout_constraintHorizontal_bias="0.004"
+                    android:id="@+id/textView_sustainability_title"
+                    android:layout_width="350dp"
+                    android:layout_height="26dp"
+                    android:text="SUSTAINABILITY FOR STUDENTS OF CASCADIA"
+                    android:textAlignment="center"
+                    android:textSize="20sp"
+                    android:textStyle="bold"
+                    android:typeface="sans"
+                    app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent" />
 
             </androidx.constraintlayout.widget.ConstraintLayout>
         </LinearLayout>
-
     </com.google.android.material.card.MaterialCardView>
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
