+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -25,7 +25,7 @@
 
     <fragment
         android:id="@+id/nav_info"
-        android:name="edu.cascadia.mobas.campusguidebook.ui.info.InfoFragment"
+        android:name="edu.cascadia.mobas.campusguidebook.ui.Sustainability.InfoFragment"
         android:label="@string/menu_info"
         tools:layout="@layout/fragment_info" >
         <action
